Moves some classes + clean up
