+++ src/main/java/adapters/custom_game/CustomPanelConstants.java
@@ -1,4 +1,4 @@
-package adapters.custom_game.custom_game_UI_adapters;
+package adapters.custom_game;
 
 public class CustomPanelConstants {
     //credit to Seamus for thinking of storing constants this way
