+++ src/main/java/adapters/custom_game/TempMazeAdapter.java
@@ -1,4 +1,4 @@
-package adapters.custom_game.custom_game_UI_adapters;
+package adapters.custom_game;
 
 import use_cases.custom_game.custom_game_editor.TempMaze;
