+++ src/main/java/user_interface/custom_game/custom_game_panels/CustomGameMainPanel.java
@@ -1,5 +1,5 @@
 package user_interface.custom_game.custom_game_panels;
-import adapters.custom_game.custom_game_UI_adapters.CustomGameGeneralInputHandler;
+import adapters.custom_game.CustomGameGeneralInputHandler;
 import user_interface.login_leaderboard.Panel;
 
 import javax.swing.*;
