+++ src/main/java/user_interface/custom_game/custom_game_panels/ICustomGamePanel.java
@@ -1,6 +1,6 @@
 package user_interface.custom_game.custom_game_panels;
 
-import adapters.custom_game.custom_game_UI_adapters.CustomGameGeneralInputHandler;
+import adapters.custom_game.CustomGameGeneralInputHandler;
 
 import javax.swing.*;
 import java.awt.*;
@@ -11,20 +11,7 @@ import java.awt.*;
 interface ICustomGamePanel {
 
     /**
-     * Draws a button on the panel that cancels all current tasks and returns the user to the custom game main menu
-     *
-     * @param x the x positioning on the panel
-     * @param y the y positioning on the panel
-     */
-    default void returnToCustomMainButton(JComponent component, int x, int y) {
-        JButton backButton = new JButton("go back to main");
-        backButton.addActionListener(new CustomGameGeneralInputHandler("toCustomMain", new CustomGamePresenter()));
-        backButton.setBounds(x, y, 50, 30);
-        component.add(backButton);
-    }
-
-    /**
-     * Overloaded version of the above method that allows the layout to decide where the button should go
+     * Draws a button on the panel that returns the user to the custom game main menu
      */
     default void returnToCustomMainButton(JComponent component) {
         JButton backButton = new JButton("go back to main");
